REF: Create the LexiconSignature class
